+++ jOOQ-meta/src/main/java/org/jooq/meta/hsqldb/HSQLDBDatabase.java
@@ -38,8 +38,6 @@
 
 package org.jooq.meta.hsqldb;
 
-import static java.util.stream.Collectors.mapping;
-import static java.util.stream.Collectors.toList;
 import static org.jooq.Records.mapping;
 import static org.jooq.impl.DSL.decode;
 import static org.jooq.impl.DSL.falseCondition;
@@ -52,6 +50,7 @@ import static org.jooq.impl.DSL.select;
 import static org.jooq.impl.DSL.when;
 import static org.jooq.impl.SQLDataType.BIGINT;
 import static org.jooq.impl.SQLDataType.INTEGER;
+import static org.jooq.impl.SQLDataType.NUMERIC;
 import static org.jooq.impl.SQLDataType.VARCHAR;
 import static org.jooq.meta.hsqldb.information_schema.Tables.CHECK_CONSTRAINTS;
 import static org.jooq.meta.hsqldb.information_schema.Tables.COLUMNS;
@@ -67,25 +66,25 @@ import static org.jooq.meta.hsqldb.information_schema.Tables.SYSTEM_TABLES;
 import static org.jooq.meta.hsqldb.information_schema.Tables.TABLE_CONSTRAINTS;
 import static org.jooq.meta.hsqldb.information_schema.Tables.VIEWS;
 
+import java.math.BigDecimal;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.stream.Collectors;
 
 import org.jooq.DSLContext;
 import org.jooq.Field;
 import org.jooq.Record;
 import org.jooq.Record12;
 import org.jooq.Record6;
-import org.jooq.Records;
 import org.jooq.Result;
 import org.jooq.ResultQuery;
 import org.jooq.SQLDialect;
 import org.jooq.SortOrder;
 import org.jooq.TableOptions.TableType;
 import org.jooq.impl.DSL;
+import org.jooq.impl.SQLDataType;
 import org.jooq.meta.AbstractDatabase;
 import org.jooq.meta.AbstractIndexDefinition;
 import org.jooq.meta.ArrayDefinition;
@@ -391,7 +390,7 @@ public class HSQLDBDatabase extends AbstractDatabase implements ResultQueryDatab
     }
 
     @Override
-    public ResultQuery<Record12<String, String, String, String, Integer, Integer, Long, Long, Long, Long, Boolean, Long>> sequences(List<String> schemas) {
+    public ResultQuery<Record12<String, String, String, String, Integer, Integer, Long, Long, BigDecimal, BigDecimal, Boolean, Long>> sequences(List<String> schemas) {
         return create()
             .select(
                 inline(null, VARCHAR).as("catalog"),
@@ -402,8 +401,8 @@ public class HSQLDBDatabase extends AbstractDatabase implements ResultQueryDatab
                 SEQUENCES.NUMERIC_SCALE.coerce(INTEGER),
                 SEQUENCES.START_WITH.coerce(BIGINT),
                 SEQUENCES.INCREMENT.coerce(BIGINT),
-                SEQUENCES.MINIMUM_VALUE.coerce(BIGINT),
-                SEQUENCES.MAXIMUM_VALUE.coerce(BIGINT),
+                SEQUENCES.MINIMUM_VALUE.coerce(NUMERIC),
+                SEQUENCES.MAXIMUM_VALUE.coerce(NUMERIC),
                 decode(SEQUENCES.CYCLE_OPTION, inline("YES"), inline(true), inline(false)).as(SEQUENCES.CYCLE_OPTION),
                 inline(null, BIGINT).as("cache"))
             .from(SEQUENCES)
